Refactor TypeUtils
 = Move to core package. It was for historical reasons in transform because
   it was originally intended as a collection of type operations that were
   were useful in transform phases. But it's now used from everywhere.
 - Make a base class of Types, so that it does not need to be imported explicitly.
