+++ compiler/src/dotty/tools/dotc/ast/tpd.scala
@@ -5,7 +5,6 @@ package ast
 import dotty.tools.dotc.transform.{ExplicitOuter, Erasure}
 import typer.ProtoTypes
 import transform.SymUtils.*
-import transform.TypeUtils.*
 import core.*
 import Scopes.newScope
 import util.Spans.*, Types.*, Contexts.*, Constants.*, Names.*, Flags.*, NameOps.*
