+++ compiler/src/dotty/tools/dotc/core/Types.scala
@@ -43,9 +43,8 @@ import scala.annotation.internal.sharable
 import scala.annotation.threadUnsafe
 
 import dotty.tools.dotc.transform.SymUtils.*
-import dotty.tools.dotc.transform.TypeUtils.isErasedClass
 
-object Types {
+object Types extends TypeUtils {
 
   @sharable private var nextId = 0
