+++ compiler/src/dotty/tools/dotc/typer/Synthesizer.scala
@@ -12,7 +12,6 @@ import ProtoTypes.*
 import Inferencing.{fullyDefinedType, isFullyDefined}
 import ast.untpd
 import transform.SymUtils.*
-import transform.TypeUtils.*
 import transform.SyntheticMembers.*
 import util.Property
 import ast.Trees.genericEmptyTree
